Move message_bus into dispatch
