+++ src/vario/main.cpp
@@ -1,10 +1,10 @@
 #include "Arduino.h"
 #include "comms/ble.h"
 #include "comms/fanet_radio.h"
+#include "dispatch/message_bus.h"
 #include "hardware/Leaf_SPI.h"
 #include "hardware/configuration.h"
 #include "instruments/gps.h"
-#include "message_bus.h"
 #include "power.h"
 #include "taskman.h"
 #include "ui/settings/settings.h"
