--- src/vario/message_bus.cpp
@@ -1,1 +1,0 @@
-#include "message_bus.h"
